moved forceAnimationStop to the state machine
